# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Kadane Backend API
  description: Kadane Backend API 
  version: 1.0.0
servers:
  - url: https://api.kadane.xyz/v1
    description: Main production server
  - url: http://localhost:3000/v1
    description: Local development server
security: 
  -  Authorization: []
paths:
...
  /starred/problems:
    get:
      tags:
        - Starred
      summary: Get starred problems
      description: >
        Returns a list of all problems with each problem including a computed 
        boolean property "starred". This field is set to true if the user has starred 
        the problem, false otherwise.
      operationId: getStarredProblems
      responses:
        '200':
          description: A list of problems with starred flag.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProblemWithStarred'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    put:
      operationId: putStarProblem
      tags:
        - Starred
      summary: Star/unstar a problem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - problemId
              properties:
                problemId:
                  type: integer
                  format: int32
      responses:
        '200':
          description: Successfully starred/unstarred problem
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        format: int32
                      starred:
                        type: boolean
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /starred/solutions:
    get:
      tags:
        - Starred
      summary: Get starred solutions
      description: >
        Returns a list of solutions with each solution including a computed boolean 
        property "starred". This field is set to true if the solution is starred by 
        the user, false otherwise.
      operationId: getStarredSolutions
      responses:
        '200':
          description: A list of solutions with starred flag.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SolutionWithStarred'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    put:
      operationId: putStarSolution
      tags:
        - Starred
      summary: Star/unstar a solution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - solutionId
              properties:
                solutionId:
                  type: integer
                  format: int32
      responses:
        '200':
          description: Successfully starred/unstarred solution
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        format: int32
                      starred:
                        type: boolean
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /starred/submissions:
    get:
      operationId: getStarredSubmissions
      tags:
        - Starred
      summary: Get starred submissions
      responses:
        '200':
          description: Successfully retrieved starred submissions
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        stdout:
                          type: string
                        time:
                          type: string
                        memory:
                          type: integer
                          format: int32
                        stderr:
                          type: string
                        compileOutput:
                          type: string
                        message:
                          type: string
                        status:
                          type: string
                        language:
                          type: string
                        accountId:
                          type: string
                        submittedCode:
                          type: string
                          format: byte
                        submittedStdin:
                          type: string
                          format: byte
                        problemId:
                          type: integer
                          format: int32
                        createdAt:
                          type: string
                          format: date-time
                        starred:
                          type: boolean
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    put:
      operationId: putStarSubmission
      tags:
        - Starred
      summary: Star/unstar a submission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - submissionId
              properties:
                submissionId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Successfully starred/unstarred submission
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      starred:
                        type: boolean
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes: 
    Authorization:
      type: http
      scheme: bearer
      bearerFormat: jwt 
  schemas:
...
    StatusCode:
      type: string
      enum: [200, 201, 202, 204, 400, 401, 403, 404, 422, 500]
    
    # Responses
    # APIResponse is a generic response that can be used for any response
    # { "data": { any } }
...
    APIErrorData:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
          anyOf:
            - $ref: '#/components/schemas/StatusCode'
    APIError:
      allOf:
        - $ref: '#/components/schemas/APIErrorData'
        - type: object
          properties:
            error:
              type: object
              $ref: '#/components/schemas/APIErrorData'
...
    InternalServerError:
      allOf:
        - $ref: '#/components/schemas/APIError'
        - type: object
          properties:
            code:
              type: string
              enum: [500]

    # Accounts
...
    Problem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        functionName:
          type: string
        tags:
          type: array
          items:
            type: string
        code:
          type: array
          items:
            $ref: '#/components/schemas/ProblemCode'
        hints:
          type: array
          items:
            $ref: '#/components/schemas/ProblemHint'
        points:
          type: integer
        testCases:
          type: array
          items:
            $ref: '#/components/schemas/ProblemTestCase'
        starred:
          type: boolean
        solved:
          type: boolean
        totalAttempts:
          type: integer
        totalCorrect:
          type: integer
    ProblemCode:
      type: object
      properties:
        language:
          type: string
        code:
          type: string
          format: byte
    ProblemHint:
      type: object
      properties:
        description:
          type: string
          format: byte
        answer:
          type: string
          format: byte
    ProblemTestCase:
      type: object
      properties:
        description:
          type: string
        input:
          type: array
          items:
            $ref: '#/components/schemas/TestCaseInput'
        output:
          type: string
        visibility:
          type: string
          enum: [public, private]
...
    TestCaseInput:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [int, float, double, string, boolean, array of int, array of float, array of double, array of string, array of boolean]
        value:
          type: string
...
    ProblemWithStarred:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          properties:
            starred:
              type: boolean
    Solution:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: "Sample Solution"
        body:
          type: string
          example: "Solution description here."
      required:
        - id
        - title
        - body
    SolutionWithStarred:
      allOf:
        - $ref: '#/components/schemas/Solution'
        - type: object
          properties:
            starred:
              type: boolean
              example: true
  responses:
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalServerError'